Move product attribute-related logic to projectors part
