+++ src/Projector/Product/Attribute/AttributeOptionResolverInterface.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Sylake\SyliusConsumerPlugin\Attribute;
+namespace Sylake\SyliusConsumerPlugin\Projector\Product\Attribute;
 
 interface AttributeOptionResolverInterface
 {
