+++ src/Projector/Product/Attribute/CompositeAttributeProcessor.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Sylake\SyliusConsumerPlugin\Attribute;
+namespace Sylake\SyliusConsumerPlugin\Projector\Product\Attribute;
 
 use Sylake\SyliusConsumerPlugin\Model\Attribute;
 use Sylius\Component\Core\Model\ProductInterface;
