+++ src/Projector/Product/Attribute/UnitAttributeProcessor.php
@@ -2,26 +2,20 @@
 
 declare(strict_types=1);
 
-namespace Sylake\SyliusConsumerPlugin\Attribute;
+namespace Sylake\SyliusConsumerPlugin\Projector\Product\Attribute;
 
 use Sylake\SyliusConsumerPlugin\Model\Attribute;
 use Sylius\Component\Attribute\Model\AttributeValueInterface;
 use Sylius\Component\Core\Model\ProductInterface;
 
-final class StringCollectionAttributeProcessor implements AttributeProcessorInterface
+final class UnitAttributeProcessor implements AttributeProcessorInterface
 {
     /** @var AttributeValueProviderInterface */
     private $attributeValueProvider;
 
-    /** @var AttributeOptionResolverInterface */
-    private $attributeOptionResolver;
-
-    public function __construct(
-        AttributeValueProviderInterface $attributeValueProvider,
-        AttributeOptionResolverInterface $attributeOptionResolver
-    ) {
+    public function __construct(AttributeValueProviderInterface $attributeValueProvider)
+    {
         $this->attributeValueProvider = $attributeValueProvider;
-        $this->attributeOptionResolver = $attributeOptionResolver;
     }
 
     /** {@inheritdoc} */
@@ -33,7 +27,7 @@ final class StringCollectionAttributeProcessor implements AttributeProcessorInte
 
         /** @var array $data */
         $data = $attribute->data();
-        if ([] === $data) {
+        if (null === $data['amount']) {
             return [];
         }
 
@@ -43,17 +37,17 @@ final class StringCollectionAttributeProcessor implements AttributeProcessorInte
             return [];
         }
 
-        $attributeValue->setValue(implode(', ', array_map(function (string $value) use ($attribute): string {
-            return $this->attributeOptionResolver->resolve($attribute->attribute(), $attribute->locale(), $value);
-        }, $data)));
+        $attributeValue->setValue(sprintf('%s %s', $data['amount'], $data['unit']));
 
         return [$attributeValue];
     }
 
     private function supports(Attribute $attribute): bool
     {
-        return is_array($attribute->data()) && array_reduce($attribute->data(), function (bool $accumulator, $value): bool {
-            return $accumulator && is_string($value) && !empty($value);
-        }, true);
+        return is_array($attribute->data())
+            && count($attribute->data()) === 2
+            && array_key_exists('amount', $attribute->data())
+            && array_key_exists('unit', $attribute->data())
+        ;
     }
 }
