+++ src/Projector/Product/ProductSlugGeneratorInterface.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Sylake\SyliusConsumerPlugin\Projector;
+namespace Sylake\SyliusConsumerPlugin\Projector\Product;
 
 use Sylius\Component\Core\Model\ProductInterface;
